+++ protogen/workflows/v1/workflows_grpc.pb.go
@@ -23,11 +23,13 @@ const _ = grpc.SupportPackageIsVersion9
 const (
 	WorkflowsService_CreateCluster_FullMethodName            = "/workflows.v1.WorkflowsService/CreateCluster"
 	WorkflowsService_GetCluster_FullMethodName               = "/workflows.v1.WorkflowsService/GetCluster"
+	WorkflowsService_UpdateCluster_FullMethodName            = "/workflows.v1.WorkflowsService/UpdateCluster"
 	WorkflowsService_DeleteCluster_FullMethodName            = "/workflows.v1.WorkflowsService/DeleteCluster"
 	WorkflowsService_ListClusters_FullMethodName             = "/workflows.v1.WorkflowsService/ListClusters"
 	WorkflowsService_CreateWorkflow_FullMethodName           = "/workflows.v1.WorkflowsService/CreateWorkflow"
 	WorkflowsService_ListWorkflows_FullMethodName            = "/workflows.v1.WorkflowsService/ListWorkflows"
 	WorkflowsService_GetWorkflow_FullMethodName              = "/workflows.v1.WorkflowsService/GetWorkflow"
+	WorkflowsService_UpdateWorkflow_FullMethodName           = "/workflows.v1.WorkflowsService/UpdateWorkflow"
 	WorkflowsService_DeleteWorkflow_FullMethodName           = "/workflows.v1.WorkflowsService/DeleteWorkflow"
 	WorkflowsService_PublishWorkflowRelease_FullMethodName   = "/workflows.v1.WorkflowsService/PublishWorkflowRelease"
 	WorkflowsService_UnpublishWorkflowRelease_FullMethodName = "/workflows.v1.WorkflowsService/UnpublishWorkflowRelease"
@@ -43,11 +45,13 @@ const (
 type WorkflowsServiceClient interface {
 	CreateCluster(ctx context.Context, in *CreateClusterRequest, opts ...grpc.CallOption) (*Cluster, error)
 	GetCluster(ctx context.Context, in *GetClusterRequest, opts ...grpc.CallOption) (*Cluster, error)
+	UpdateCluster(ctx context.Context, in *UpdateClusterRequest, opts ...grpc.CallOption) (*Cluster, error)
 	DeleteCluster(ctx context.Context, in *DeleteClusterRequest, opts ...grpc.CallOption) (*DeleteClusterResponse, error)
 	ListClusters(ctx context.Context, in *ListClustersRequest, opts ...grpc.CallOption) (*ListClustersResponse, error)
 	CreateWorkflow(ctx context.Context, in *CreateWorkflowRequest, opts ...grpc.CallOption) (*Workflow, error)
 	ListWorkflows(ctx context.Context, in *ListWorkflowsRequest, opts ...grpc.CallOption) (*ListWorkflowsResponse, error)
 	GetWorkflow(ctx context.Context, in *GetWorkflowRequest, opts ...grpc.CallOption) (*Workflow, error)
+	UpdateWorkflow(ctx context.Context, in *UpdateWorkflowRequest, opts ...grpc.CallOption) (*Workflow, error)
 	DeleteWorkflow(ctx context.Context, in *DeleteWorkflowRequest, opts ...grpc.CallOption) (*DeleteWorkflowResponse, error)
 	PublishWorkflowRelease(ctx context.Context, in *PublishWorkflowReleaseRequest, opts ...grpc.CallOption) (*WorkflowRelease, error)
 	UnpublishWorkflowRelease(ctx context.Context, in *UnpublishWorkflowReleaseRequest, opts ...grpc.CallOption) (*UnpublishWorkflowReleaseResponse, error)
@@ -77,6 +81,16 @@ func (c *workflowsServiceClient) GetCluster(ctx context.Context, in *GetClusterR
 	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Cluster)
 	err := c.cc.Invoke(ctx, WorkflowsService_GetCluster_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *workflowsServiceClient) UpdateCluster(ctx context.Context, in *UpdateClusterRequest, opts ...grpc.CallOption) (*Cluster, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(Cluster)
+	err := c.cc.Invoke(ctx, WorkflowsService_UpdateCluster_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -127,6 +141,16 @@ func (c *workflowsServiceClient) GetWorkflow(ctx context.Context, in *GetWorkflo
 	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Workflow)
 	err := c.cc.Invoke(ctx, WorkflowsService_GetWorkflow_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *workflowsServiceClient) UpdateWorkflow(ctx context.Context, in *UpdateWorkflowRequest, opts ...grpc.CallOption) (*Workflow, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(Workflow)
+	err := c.cc.Invoke(ctx, WorkflowsService_UpdateWorkflow_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -191,11 +215,13 @@ func (c *workflowsServiceClient) UndeployWorkflowRelease(ctx context.Context, in
 type WorkflowsServiceServer interface {
 	CreateCluster(context.Context, *CreateClusterRequest) (*Cluster, error)
 	GetCluster(context.Context, *GetClusterRequest) (*Cluster, error)
+	UpdateCluster(context.Context, *UpdateClusterRequest) (*Cluster, error)
 	DeleteCluster(context.Context, *DeleteClusterRequest) (*DeleteClusterResponse, error)
 	ListClusters(context.Context, *ListClustersRequest) (*ListClustersResponse, error)
 	CreateWorkflow(context.Context, *CreateWorkflowRequest) (*Workflow, error)
 	ListWorkflows(context.Context, *ListWorkflowsRequest) (*ListWorkflowsResponse, error)
 	GetWorkflow(context.Context, *GetWorkflowRequest) (*Workflow, error)
+	UpdateWorkflow(context.Context, *UpdateWorkflowRequest) (*Workflow, error)
 	DeleteWorkflow(context.Context, *DeleteWorkflowRequest) (*DeleteWorkflowResponse, error)
 	PublishWorkflowRelease(context.Context, *PublishWorkflowReleaseRequest) (*WorkflowRelease, error)
 	UnpublishWorkflowRelease(context.Context, *UnpublishWorkflowReleaseRequest) (*UnpublishWorkflowReleaseResponse, error)
@@ -217,6 +243,9 @@ func (UnimplementedWorkflowsServiceServer) CreateCluster(context.Context, *Creat
 func (UnimplementedWorkflowsServiceServer) GetCluster(context.Context, *GetClusterRequest) (*Cluster, error) {
 	return nil, status.Error(codes.Unimplemented, "method GetCluster not implemented")
 }
+func (UnimplementedWorkflowsServiceServer) UpdateCluster(context.Context, *UpdateClusterRequest) (*Cluster, error) {
+	return nil, status.Error(codes.Unimplemented, "method UpdateCluster not implemented")
+}
 func (UnimplementedWorkflowsServiceServer) DeleteCluster(context.Context, *DeleteClusterRequest) (*DeleteClusterResponse, error) {
 	return nil, status.Error(codes.Unimplemented, "method DeleteCluster not implemented")
 }
@@ -231,6 +260,9 @@ func (UnimplementedWorkflowsServiceServer) ListWorkflows(context.Context, *ListW
 }
 func (UnimplementedWorkflowsServiceServer) GetWorkflow(context.Context, *GetWorkflowRequest) (*Workflow, error) {
 	return nil, status.Error(codes.Unimplemented, "method GetWorkflow not implemented")
+}
+func (UnimplementedWorkflowsServiceServer) UpdateWorkflow(context.Context, *UpdateWorkflowRequest) (*Workflow, error) {
+	return nil, status.Error(codes.Unimplemented, "method UpdateWorkflow not implemented")
 }
 func (UnimplementedWorkflowsServiceServer) DeleteWorkflow(context.Context, *DeleteWorkflowRequest) (*DeleteWorkflowResponse, error) {
 	return nil, status.Error(codes.Unimplemented, "method DeleteWorkflow not implemented")
@@ -300,6 +332,24 @@ func _WorkflowsService_GetCluster_Handler(srv interface{}, ctx context.Context, 
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(WorkflowsServiceServer).GetCluster(ctx, req.(*GetClusterRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _WorkflowsService_UpdateCluster_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(UpdateClusterRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(WorkflowsServiceServer).UpdateCluster(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: WorkflowsService_UpdateCluster_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(WorkflowsServiceServer).UpdateCluster(ctx, req.(*UpdateClusterRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -390,6 +440,24 @@ func _WorkflowsService_GetWorkflow_Handler(srv interface{}, ctx context.Context,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(WorkflowsServiceServer).GetWorkflow(ctx, req.(*GetWorkflowRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _WorkflowsService_UpdateWorkflow_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(UpdateWorkflowRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(WorkflowsServiceServer).UpdateWorkflow(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: WorkflowsService_UpdateWorkflow_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(WorkflowsServiceServer).UpdateWorkflow(ctx, req.(*UpdateWorkflowRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -500,6 +568,10 @@ var WorkflowsService_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _WorkflowsService_GetCluster_Handler,
 		},
 		{
+			MethodName: "UpdateCluster",
+			Handler:    _WorkflowsService_UpdateCluster_Handler,
+		},
+		{
 			MethodName: "DeleteCluster",
 			Handler:    _WorkflowsService_DeleteCluster_Handler,
 		},
@@ -518,6 +590,10 @@ var WorkflowsService_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "GetWorkflow",
 			Handler:    _WorkflowsService_GetWorkflow_Handler,
+		},
+		{
+			MethodName: "UpdateWorkflow",
+			Handler:    _WorkflowsService_UpdateWorkflow_Handler,
 		},
 		{
 			MethodName: "DeleteWorkflow",