+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2026-07-01
+
+### Added
+
+- `workflows`: Added `client.Clusters.Update`.
+- `workflows`: Added `client.Workflows.Update`.
+- `workflows`: Added cluster descriptions.
+- `workflows`: Added deployed cluster details on workflow releases.
+
 ## [0.9.0] - 2026-06-25
 
 ### Added
@@ -139,7 +148,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for Tilebox Observability, including logging and tracing helpers.
 - Added examples for using the library.
 
-[Unreleased]: https://github.com/tilebox/tilebox-go/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-go/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/tilebox/tilebox-go/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/tilebox/tilebox-go/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/tilebox/tilebox-go/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/tilebox/tilebox-go/compare/v0.7.0...v0.7.1